Data directory size variable changed to long

+++ src/main/offline/io/stargate/sgv2/jsonapi/service/cqldriver/sstablewriter/OfflineWriterSessionStatus.java
@@ -6,6 +6,6 @@ public record OfflineWriterSessionStatus(
     String tableName,
     String ssTableOutputDirectory,
     int fileWriterBufferSizeInMB,
-    int dataDirectorySizeInBytes,
+    long dataDirectorySizeInBytes,
     int insertsSucceeded,
     int insertsFailed) {}
